Basic implementation of add
All it does is append the fetched document at the end of the file,
regardless of whether or not it's already present in the file.

In the future, the function will first check if the document is already
present. If so, it removes it (first making a backup) and then fetches
a new version that it appends at then end.